+++ src/test/kotlin/HuffmanTest.kt
@@ -18,10 +18,10 @@ class HuffmanTest {
         )
 
         val tokens: MutableList<LZ77Token> = mutableListOf()
-        input.forEach { tokens.add(LZ77Literal(it)) }
+        input.forEach { tokens.add(LZ77Literal(it.toUByte())) }
         tokens.add(LZ77Repeat(24, 12))
         tokens.add(LZ77Repeat(24, 12))
-        tokens.add(LZ77Literal(0x77))
+        tokens.add(LZ77Literal(0x77.toUByte()))
 
         val huffman = HuffmanCompressor()
         val output = huffman.encode(tokens)
@@ -34,7 +34,7 @@ class HuffmanTest {
         val inputBytes = inputFile.readBytes()
 
         val tokens: MutableList<LZ77Token> = mutableListOf()
-        inputBytes.forEach { tokens.add(LZ77Literal(it)) }
+        inputBytes.forEach { tokens.add(LZ77Literal(it.toUByte())) }
 
         val huffman = HuffmanCompressor()
         val map = huffman.computeFrequencies(tokens)
@@ -42,13 +42,13 @@ class HuffmanTest {
             println("Key: ${token.key.toInt().toChar()}, Value: ${token.value}")
         }
 
-        assert(map['B'.code.toByte()] == 1)
-        assert(map['l'.code.toByte()] == 5)
-        assert(map['a'.code.toByte()] == 5)
-        assert(map['h'.code.toByte()] == 5)
-        assert(map[' '.code.toByte()] == 4)
-        assert(map['b'.code.toByte()] == 4)
-        assert(map['!'.code.toByte()] == 1)
+        assert(map['B'.code.toUByte()] == 1)
+        assert(map['l'.code.toUByte()] == 5)
+        assert(map['a'.code.toUByte()] == 5)
+        assert(map['h'.code.toUByte()] == 5)
+        assert(map[' '.code.toUByte()] == 4)
+        assert(map['b'.code.toUByte()] == 4)
+        assert(map['!'.code.toUByte()] == 1)
     }
 
     @Test
@@ -58,7 +58,7 @@ class HuffmanTest {
         val inputBytes = inputFile.readBytes()
 
         val tokens: MutableList<LZ77Token> = mutableListOf()
-        inputBytes.forEach { tokens.add(LZ77Literal(it)) }
+        inputBytes.forEach { tokens.add(LZ77Literal(it.toUByte())) }
 
         val huffman = HuffmanCompressor()
         val map = huffman.computeFrequencies(tokens)
@@ -67,19 +67,19 @@ class HuffmanTest {
         assert(tree.weight == 15)
         // a
         assert(tree.left.weight == 6)
-        assert((tree.left as LeafNode).symbol == 'a'.code.toByte())
+        assert((tree.left as LeafNode).symbol == 'a'.code.toUByte())
 
         assert(tree.right.weight == 9)
         // b
         assert((tree.right as CompositeNode).left.weight == 4)
-        assert(((tree.right as CompositeNode).left as LeafNode).symbol == 'b'.code.toByte())
+        assert(((tree.right as CompositeNode).left as LeafNode).symbol == 'b'.code.toUByte())
 
         assert((tree.right as CompositeNode).right.weight == 5)
         // c
         assert(((tree.right as CompositeNode).right as CompositeNode).left.weight == 2)
-        assert((((tree.right as CompositeNode).right as CompositeNode).left as LeafNode).symbol == 'c'.code.toByte())
+        assert((((tree.right as CompositeNode).right as CompositeNode).left as LeafNode).symbol == 'c'.code.toUByte())
         // d
         assert(((tree.right as CompositeNode).right as CompositeNode).right.weight == 3)
-        assert((((tree.right as CompositeNode).right as CompositeNode).right as LeafNode).symbol == 'd'.code.toByte())
+        assert((((tree.right as CompositeNode).right as CompositeNode).right as LeafNode).symbol == 'd'.code.toUByte())
     }
 }
